// components/controller/src/extensions/soviet-robot/infrastructure/database/soviet-robot.database.types.ts
/**
 * This file was generated by kysely-codegen.
 * Please do not edit it manually.
 */

import type { ColumnType } from "kysely";

export type Generated<T> = T extends ColumnType<infer S, infer I, infer U>
  ? ColumnType<S, I | undefined, U>
  : ColumnType<T, T | undefined, T>;

export type Json = JsonValue;

export type JsonArray = JsonValue[];

export type JsonObject = {
  [x: string]: JsonValue | undefined;
};

export type JsonPrimitive = boolean | number | string | null;

export type JsonValue = JsonArray | JsonObject | JsonPrimitive;

export type Timestamp = ColumnType<Date, Date | string, Date | string>;

export interface SovietRobotDecisions {
  attempts: Generated<number>;
  coopname: string;
  created_at: Generated<Timestamp>;
  decision_hash: string;
  decision_id: number;
  decision_type: string;
  id: Generated<string>;
  last_error: string | null;
  next_attempt_at: Timestamp | null;
  protocol_hash: string | null;
  stage: Generated<string>;
  tx_hashes: Generated<Json>;
  updated_at: Generated<Timestamp>;
  username: string;
  votes: Generated<Json>;
  waiting_for: Generated<Json>;
}

export interface SovietRobotKeys {
  coopname: string;
  created_at: Generated<Timestamp>;
  encrypted_wif: string;
  id: Generated<string>;
  member: string;
  permission_name: string;
  public_key: string;
  updated_at: Generated<Timestamp>;
}

export interface DB {
  soviet_robot_decisions: SovietRobotDecisions;
  soviet_robot_keys: SovietRobotKeys;
}